# 4-doc-doc-testing-load-testing/workshop/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Workshop 4 API",
    "description": "API",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/todo": {
      "get": {
        "tags": [
          "todo"
        ],
        "summary": "Get all todo",
        "description": "Get all todo",
        "operationId": "getTodo",
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/todos"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "tags": [
          "todo"
        ],
        "summary": "Create a todo",
        "description": "Create a todo",
        "operationId": "createTodo",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Todo to create",
            "required": true,
            "schema": {
              "type": "string",
              "required": ["text"],
              "properties": {
                "text": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/todo"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/todo/{id}": {
      "patch": {
        "tags": [
          "todo"
        ],
        "summary": "Update a todo",
        "description": "Update a todo",
        "operationId": "updateTodo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of todo to update",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/todo"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  },
  "definitions": {
    "todo": {
      "type": "object",
      "required": ["text"],
      "properties": {
        "_id": {
          "type": "string"
        },
        "text": {
          "type": "string"
        },
        "done": {
          "type": "boolean"
        },
        "__v": {
          "type": "integer"
        }
      }
    },
    "todos": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/todo"
      }
    }
  }
}